+++ src/test/scala/io/boson/injectorAPITests.scala
@@ -25,293 +25,22 @@ class injectorAPITests extends FunSuite {
   val validatedByteArray: Array[Byte] = arr.encodeToBarray()
   val validatedByteArrayObj: Array[Byte] = bsonEvent.encodeToBarray()
 
-  test("Inject API Double => Double") {
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false)
-    val newFridgeSerialCode: Double = 1000.0
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "fanVelocity.first"
-    val boson: Boson = Boson.injector(expression, (in: Double) => newFridgeSerialCode)
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(List(1000.0), future.join().getValue )
-  }
-  test("Inject API String => String") {
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the")
-    val newFridgeSerialCode: String = " what?"
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "string.first"
-    val boson: Boson = Boson.injector(expression, (in: String) => in.concat(newFridgeSerialCode))
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(List("the what?").head, new String(future.join().getValue.asInstanceOf[List[Array[Byte]]].head) )
-  }
-  test("Inject API Map => Map") {
-    val bAux: BsonObject = new BsonObject().put("damnnn", "DAMMN")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the").put("bson", bAux)
-
-    val newFridgeSerialCode: String = " what?"
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: Map[String, Any]) => in.+(("WHAT!!!", 10)))
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(Map("damnnn" -> "DAMMN", "WHAT!!!" -> 10))),future.join() )
-  }
-  test("Inject API Map => Map 1") {
-    val bAux: BsonObject = new BsonObject().put("damnnn", "DAMMN")
-    val bAux1: BsonObject = new BsonObject().put("DAMNNNNN", "damnn")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the").put("bson", bAux)
-
-    val newFridgeSerialCode: Map[String, Any] = Mapper.convert(bAux1).asInstanceOf[ Map[String, Any]]
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: Map[String, Any]) => newFridgeSerialCode)
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(Map("DAMNNNNN" -> "damnn"))),future.join() )
-  }
-  test("Inject API List => List") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode: String = "MAIS EU"
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: List[Any]) => {
-      val s: List[Any] = in.:+(newFridgeSerialCode)
-      s})
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(12, "sddd", "MAIS EU")),future.join() )
-  }
-  test("Inject API List => List 1") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bAux1: BsonArray = new BsonArray().add("sddd").add(12)
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode:List[Any] = Mapper.convertBsonArray(bAux1)
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: List[Any]) => newFridgeSerialCode)
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List("sddd", 12)),future.join() )
-  }
-  test("Inject API Int => Int") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("int", 10).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode: Int = 2
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "int.first"
-    val boson: Boson = Boson.injector(expression, (in: Int) => in*newFridgeSerialCode)
-    val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(20)),future.join() )
-  }
-  test("Inject API Long => Long") {
+ /* test("MORE KEYS"){
     val bAux: BsonArray = new BsonArray().add(12).add("sddd")
     val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("long", 9L).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
 
     val newFridgeSerialCode: Long = 2
-    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray
-    val expression = "long.first"
+    val validBsonArray: Array[Byte] = bsonEvent.encodeToBarray()
+    val expression = "ola.[2 to 2].segunda.[0 until end].quarta.quinta.terceira.first..first"
     val boson: Boson = Boson.injector(expression, (in: Long) => in*newFridgeSerialCode)
     val result: CompletableFuture[Array[Byte]] = boson.go(validBsonArray)
 
-    // apply an extractor to get the new serial code as above.
     val resultValue: Array[Byte] = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
+    /*val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
     val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
+    boson1.go(resultValue)*/
 
-    assertEquals(BsSeq(List(18)),future.join() )
-  }
+    assertEquals(BsSeq(List(18)),resultValue )
 
-  test("Inject API Double => Double ByteBuffer") {
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false)
-    val newFridgeSerialCode: Double = 1000.0
-    val validBsonArray: ByteBuffer = bsonEvent.encode().getByteBuf.nioBuffer()
-    val expression = "fanVelocity.first"
-    val boson: Boson = Boson.injector(expression, (in: Double) => newFridgeSerialCode)
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(List(1000.0), future.join().getValue )
-  }
-  test("Inject API String => String ByteBuffer") {
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the")
-    val newFridgeSerialCode: String = " what?"
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "string.first"
-    val boson: Boson = Boson.injector(expression, (in: String) => in.concat(newFridgeSerialCode))
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue:ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(List("the what?").head, new String(future.join().getValue.asInstanceOf[List[Array[Byte]]].head) )
-  }
-  test("Inject API Map => Map ByteBuffer") {
-    val bAux: BsonObject = new BsonObject().put("damnnn", "DAMMN")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the").put("bson", bAux)
-
-    val newFridgeSerialCode: String = " what?"
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: Map[String, Any]) => in.+(("WHAT!!!", 10)))
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(Map("damnnn" -> "DAMMN", "WHAT!!!" -> 10))),future.join() )
-  }
-  test("Inject API Map => Map 1 ByteBuffer") {
-    val bAux: BsonObject = new BsonObject().put("damnnn", "DAMMN")
-    val bAux1: BsonObject = new BsonObject().put("DAMNNNNN", "damnn")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", "the").put("bson", bAux)
-
-    val newFridgeSerialCode: Map[String, Any] = Mapper.convert(bAux1).asInstanceOf[ Map[String, Any]]
-    val validBsonArray:ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: Map[String, Any]) => newFridgeSerialCode)
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue:ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(Map("DAMNNNNN" -> "damnn"))),future.join() )
-  }
-  test("Inject API List => List ByteBuffer") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode: String = "MAIS EU"
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: List[Any]) => {
-      val s: List[Any] = in.:+(newFridgeSerialCode)
-      s})
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(12, "sddd", "MAIS EU")),future.join() )
-  }
-  test("Inject API List => List 1 ByteBuffer") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bAux1: BsonArray = new BsonArray().add("sddd").add(12)
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode:List[Any] = Mapper.convertBsonArray(bAux1)
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "bson.first"
-    val boson: Boson = Boson.injector(expression, (in: List[Any]) => newFridgeSerialCode)
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List("sddd", 12)),future.join() )
-  }
-  test("Inject API Int => Int ByteBuffer") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("int", 10).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode: Int = 2
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "int.first"
-    val boson: Boson = Boson.injector(expression, (in: Int) => in*newFridgeSerialCode)
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(20)),future.join() )
-  }
-  test("Inject API Long => Long ByteBuffer") {
-    val bAux: BsonArray = new BsonArray().add(12).add("sddd")
-    val bsonEvent: BsonObject = new BsonObject().put("fridgeTemp", 5.2f).put("long", 9L).put("fanVelocity", 20.5).put("doorOpen", false).put("string", 1).put("bson", bAux)
-
-    val newFridgeSerialCode: Long = 2
-    val validBsonArray: ByteBuffer = bsonEvent.encode.getByteBuf.nioBuffer()
-    val expression = "long.first"
-    val boson: Boson = Boson.injector(expression, (in: Long) => in*newFridgeSerialCode)
-    val result: CompletableFuture[ByteBuffer] = boson.go(validBsonArray)
-
-    // apply an extractor to get the new serial code as above.
-    val resultValue: ByteBuffer = result.join()
-    val future: CompletableFuture[BsValue] = new CompletableFuture[BsValue]()
-    val boson1: Boson = Boson.extractor(expression, (in: BsValue) => future.complete(in))
-    boson1.go(resultValue)
-
-    assertEquals(BsSeq(List(18)),future.join() )
-  }
+  }*/
 }
